+++ packages/adapter-cloudflare/index.js
@@ -1,57 +1,129 @@
-import { existsSync, writeFileSync } from 'node:fs';
-import * as path from 'node:path';
+import { VERSION } from '@sveltejs/kit';
+import { copyFileSync, existsSync, writeFileSync } from 'node:fs';
+import path from 'node:path';
+import process from 'node:process';
 import { fileURLToPath } from 'node:url';
-import { getPlatformProxy } from 'wrangler';
+import { getPlatformProxy, unstable_readConfig } from 'wrangler';
+import { is_building_for_cloudflare_pages, validate_worker_settings } from './utils.js';
+
+const name = '@sveltejs/adapter-cloudflare';
+const [kit_major, kit_minor] = VERSION.split('.');
 
 /** @type {import('./index.js').default} */
 export default function (options = {}) {
 	return {
-		name: '@sveltejs/adapter-cloudflare',
+		name,
 		async adapt(builder) {
 			if (existsSync('_routes.json')) {
 				throw new Error(
-					'Cloudflare routes should be configured in svelte.config.js rather than _routes.json'
+					"Cloudflare Pages' _routes.json should be configured in svelte.config.js. See https://svelte.dev/docs/kit/adapter-cloudflare#Options-routes"
 				);
 			}
 
+			if (existsSync(`${builder.config.kit.files.assets}/_headers`)) {
+				throw new Error(
+					`The _headers file should be placed in the project root rather than the ${builder.config.kit.files.assets} directory`
+				);
+			}
+
+			if (existsSync(`${builder.config.kit.files.assets}/_redirects`)) {
+				throw new Error(
+					`The _redirects file should be placed in the project root rather than the ${builder.config.kit.files.assets} directory`
+				);
+			}
+
+			const wrangler_config = validate_wrangler_config(options.config);
+
+			const building_for_cloudflare_pages = is_building_for_cloudflare_pages(wrangler_config);
+
+			let dest = builder.getBuildDirectory('cloudflare');
+			let worker_dest = `${dest}/_worker.js`;
+			let assets_binding = 'ASSETS';
+
+			if (building_for_cloudflare_pages) {
+				if (wrangler_config.pages_build_output_dir) {
+					dest = wrangler_config.pages_build_output_dir;
+					worker_dest = `${dest}/_worker.js`;
+				}
+			} else {
+				if (wrangler_config.main) {
+					worker_dest = wrangler_config.main;
+				}
+				if (wrangler_config.assets?.directory) {
+					// wrangler doesn't resolve `assets.directory` to an absolute path unlike
+					// `main` and `pages_build_output_dir` so we need to do it ourselves here
+					const parent_dir = wrangler_config.configPath
+						? path.dirname(path.resolve(wrangler_config.configPath))
+						: process.cwd();
+					dest = path.resolve(parent_dir, wrangler_config.assets.directory);
+				}
+				if (wrangler_config.assets?.binding) {
+					assets_binding = wrangler_config.assets.binding;
+				}
+			}
+
 			const files = fileURLToPath(new URL('./files', import.meta.url).href);
-			const dest = builder.getBuildDirectory('cloudflare');
 			const tmp = builder.getBuildDirectory('cloudflare-tmp');
 
 			builder.rimraf(dest);
-			builder.rimraf(tmp);
+			builder.rimraf(worker_dest);
 
 			builder.mkdirp(dest);
 			builder.mkdirp(tmp);
 
-			// generate plaintext 404.html first which can then be overridden by prerendering, if the user defined such a page
-			const fallback = path.join(dest, '404.html');
-			if (options.fallback === 'spa') {
-				await builder.generateFallback(fallback);
-			} else {
-				writeFileSync(fallback, 'Not Found');
+			// client assets and prerendered pages
+			const assets_dest = `${dest}${builder.config.kit.paths.base}`;
+			builder.mkdirp(assets_dest);
+			if (
+				building_for_cloudflare_pages ||
+				wrangler_config.assets?.not_found_handling === '404-page'
+			) {
+				// generate plaintext 404.html first which can then be overridden by prerendering, if the user defined such a page
+				const fallback = path.join(assets_dest, '404.html');
+				if (options.fallback === 'spa') {
+					await builder.generateFallback(fallback);
+				} else {
+					writeFileSync(fallback, 'Not Found');
+				}
+			}
+			const client_assets = builder.writeClient(assets_dest);
+			builder.writePrerendered(assets_dest);
+			if (
+				!building_for_cloudflare_pages &&
+				wrangler_config.assets?.not_found_handling === 'single-page-application'
+			) {
+				await builder.generateFallback(path.join(assets_dest, 'index.html'));
 			}
 
-			const dest_dir = `${dest}${builder.config.kit.paths.base}`;
-			const written_files = builder.writeClient(dest_dir);
-			builder.writePrerendered(dest_dir);
-
-			const relativePath = path.posix.relative(dest, builder.getServerDirectory());
-
+			// worker
+			const worker_dest_dir = path.dirname(worker_dest);
 			writeFileSync(
 				`${tmp}/manifest.js`,
-				`export const manifest = ${builder.generateManifest({ relativePath })};\n\n` +
+				`export const manifest = ${builder.generateManifest({ relativePath: path.posix.relative(tmp, builder.getServerDirectory()) })};\n\n` +
 					`export const prerendered = new Set(${JSON.stringify(builder.prerendered.paths)});\n\n` +
 					`export const base_path = ${JSON.stringify(builder.config.kit.paths.base)};\n`
 			);
+			builder.copy(`${files}/worker.js`, worker_dest, {
+				replace: {
+					// the paths returned by the Wrangler config might be Windows paths,
+					// so we need to convert them to POSIX paths or else the backslashes
+					// will be interpreted as escape characters and create an incorrect import path
+					SERVER: `${posixify(path.relative(worker_dest_dir, builder.getServerDirectory()))}/index.js`,
+					MANIFEST: `${posixify(path.relative(worker_dest_dir, tmp))}/manifest.js`,
+					ASSETS: assets_binding
+				}
+			});
 
-			writeFileSync(
-				`${dest}/_routes.json`,
-				JSON.stringify(get_routes_json(builder, written_files, options.routes ?? {}), null, '\t')
-			);
-
+			// _headers
+			if (existsSync('_headers')) {
+				copyFileSync('_headers', `${dest}/_headers`);
+			}
 			writeFileSync(`${dest}/_headers`, generate_headers(builder.getAppPath()), { flag: 'a' });
 
+			// _redirects
+			if (existsSync('_redirects')) {
+				copyFileSync('_redirects', `${dest}/_redirects`);
+			}
 			if (builder.prerendered.redirects.size > 0) {
 				writeFileSync(`${dest}/_redirects`, generate_redirects(builder.prerendered.redirects), {
 					flag: 'a'
@@ -60,12 +132,12 @@ export default function (options = {}) {
 
 			writeFileSync(`${dest}/.assetsignore`, generate_assetsignore(), { flag: 'a' });
 
-			builder.copy(`${files}/worker.js`, `${dest}/_worker.js`, {
-				replace: {
-					SERVER: `${relativePath}/index.js`,
-					MANIFEST: `${path.posix.relative(dest, tmp)}/manifest.js`
-				}
-			});
+			if (building_for_cloudflare_pages) {
+				writeFileSync(
+					`${dest}/_routes.json`,
+					JSON.stringify(get_routes_json(builder, client_assets, options.routes ?? {}), null, '\t')
+				);
+			}
 		},
 		emulate() {
 			// we want to invoke `getPlatformProxy` only once, but await it only when it is accessed.
@@ -74,7 +146,8 @@ export default function (options = {}) {
 				const proxy = await getPlatformProxy(options.platformProxy);
 				const platform = /** @type {App.Platform} */ ({
 					env: proxy.env,
-					context: proxy.ctx,
+					ctx: proxy.ctx,
+					context: proxy.ctx, // deprecated in favor of ctx
 					caches: proxy.caches,
 					cf: proxy.cf
 				});
@@ -100,6 +173,18 @@ export default function (options = {}) {
 					return prerender ? emulated.prerender_platform : emulated.platform;
 				}
 			};
+		},
+		supports: {
+			read: ({ route }) => {
+				// TODO bump peer dep in next adapter major to simplify this
+				if (kit_major === '2' && kit_minor < '25') {
+					throw new Error(
+						`${name}: Cannot use \`read\` from \`$app/server\` in route \`${route.id}\` when using SvelteKit < 2.25.0`
+					);
+				}
+
+				return true;
+			}
 		}
 	};
 }
@@ -152,7 +237,7 @@ function get_routes_json(builder, assets, { include = ['/*'], exclude = ['<all>'
 
 	const excess = include.length + exclude.length - 100;
 	if (excess > 0) {
-		const message = `Function includes/excludes exceeds _routes.json limits (see https://developers.cloudflare.com/pages/platform/functions/routing/#limits). Dropping ${excess} exclude rules — this will cause unnecessary function invocations.`;
+		const message = `Cloudflare Pages Functions' includes/excludes exceeds _routes.json limits (see https://developers.cloudflare.com/pages/platform/functions/routing/#limits). Dropping ${excess} exclude rules — this will cause unnecessary function invocations.`;
 		builder.log.warn(message);
 
 		exclude.length -= excess;
@@ -202,4 +287,24 @@ _routes.json
 _headers
 _redirects
 `;
+}
+
+/**
+ * @param {string} config_file
+ * @returns {import('wrangler').Unstable_Config}
+ */
+function validate_wrangler_config(config_file = undefined) {
+	const wrangler_config = unstable_readConfig({ config: config_file });
+
+	if (!is_building_for_cloudflare_pages(wrangler_config)) {
+		// probably deploying to Cloudflare Workers
+		validate_worker_settings(wrangler_config);
+	}
+
+	return wrangler_config;
+}
+
+/** @param {string} str */
+function posixify(str) {
+	return str.replace(/\\/g, '/');
 }