+++ packages/kit/src/core/config/index.spec.js
@@ -76,7 +76,11 @@ const get_defaults = (prefix = '') => ({
 			publicPrefix: 'PUBLIC_',
 			privatePrefix: ''
 		},
+		experimental: {
+			remoteFunctions: false
+		},
 		files: {
+			src: join(prefix, 'src'),
 			assets: join(prefix, 'static'),
 			hooks: {
 				client: join(prefix, 'src/hooks.client'),
@@ -365,6 +369,30 @@ test('load default config (esm)', async () => {
 	expect(config).toEqual(defaults);
 });
 
+test('load default config (esm) with .ts extensions', async () => {
+	const cwd = join(__dirname, 'fixtures/typescript');
+
+	const config = await load_config({ cwd });
+	remove_keys(config, ([, v]) => typeof v === 'function');
+
+	const defaults = get_defaults(cwd + '/');
+	defaults.kit.version.name = config.kit.version.name;
+
+	expect(config).toEqual(defaults);
+});
+
+test('load .js config when both .js and .ts configs are present', async () => {
+	const cwd = join(__dirname, 'fixtures/multiple');
+
+	const config = await load_config({ cwd });
+	remove_keys(config, ([, v]) => typeof v === 'function');
+
+	const defaults = get_defaults(cwd + '/');
+	defaults.kit.version.name = config.kit.version.name;
+
+	expect(config).toEqual(defaults);
+});
+
 test('errors on loading config with incorrect default export', async () => {
 	let message = null;
 
@@ -377,6 +405,18 @@ test('errors on loading config with incorrect default export', async () => {
 
 	assert.equal(
 		message,
-		'svelte.config.js must have a configuration object as its default export. See https://svelte.dev/docs/kit/configuration'
+		'The Svelte config file must have a configuration object as its default export. See https://svelte.dev/docs/kit/configuration'
 	);
+});
+
+test('uses src prefix for other kit.files options', async () => {
+	const cwd = join(__dirname, 'fixtures/custom-src');
+
+	const config = await load_config({ cwd });
+	remove_keys(config, ([, v]) => typeof v === 'function');
+
+	const defaults = get_defaults(cwd + '/');
+	defaults.kit.version.name = config.kit.version.name;
+
+	expect(config.kit.files.lib).toEqual(join(cwd, 'source/lib'));
 });