+++ packages/kit/src/utils/filesystem.js
@@ -105,7 +105,7 @@ export function walk(cwd, dirs = false) {
 		}
 	}
 
-	return walk_dir(''), all_files;
+	return (walk_dir(''), all_files);
 }
 
 /** @param {string} str */